+++ contracts/interfaces/iERC20.sol
@@ -6,7 +6,7 @@ interface iERC20 {
 
     function symbol() external view returns (string memory);
 
-    function decimals() external view returns (uint256);
+    function decimals() external view returns (uint8);
 
     function totalSupply() external view returns (uint256);
